requirements.txt
bcrypt
Flask
flask_cors
google_generativeai
numpy
opencv_python
Pillow
pytesseract
python_dotenv
pyzbar
requests
supabase